[refactor] InputBox: InputTextarea 리네이밍, 모드 추가
